% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMatch_annotation.R
\name{ClusterMatch_annotation}
\alias{ClusterMatch_annotation}
\title{Annotate the query data based on the reference data labels}
\usage{
ClusterMatch_annotation(
  reference_matrix,
  query_matrix,
  reference_label,
  que.res = 2,
  ref.norm = TRUE,
  que.norm = TRUE,
  dim.pca = 50,
  dim.cca = 30,
  min.cluster.cells = 20,
  random_PCC = 0
)
}
\arguments{
\item{reference_matrix}{the gene expression matrix of reference}

\item{query_matrix}{the gene expression matrix of query}

\item{reference_label}{reference data labels}

\item{que.res}{the query clustering resolution, default 2}

\item{ref.norm}{whether reference is normalized}

\item{que.norm}{whether query is normalized}

\item{dim.pca}{number of PCs to calculate (50 by default)}

\item{dim.cca}{number of canonical vectors to calculate (30 by default)}

\item{min.cluster.cells}{the minimum number of cells in a cluster}

\item{random_PCC}{the minimum value of random noise}
}
\value{
a list containing the following elements:
\itemize{
\item D1_reference: SeuratObject of reference
\item D2_query: SeuratObject of query
\item Matching_matrix: Matching matrix of cell types and clusters
}
}
\description{
Annotate the query data based on the reference data labels
}
