% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMatch_integration.R
\name{ClusterMatch_integration}
\alias{ClusterMatch_integration}
\title{Integrate scRNA-seq datasets}
\usage{
ClusterMatch_integration(
  reference_matrix,
  query_matrix,
  ref.res,
  que.res,
  ref.norm = TRUE,
  que.norm = TRUE,
  dim.pca = 50,
  dim.cca = 30,
  min.cluster.cells = 20,
  random_PCC = 0,
  distance_diff = 0,
  distance_same = 0
)
}
\arguments{
\item{reference_matrix}{the gene expression matrix of dataset 1}

\item{query_matrix}{the gene expression matrix of dataset 2}

\item{ref.res}{the optimal clustering resolution of dataset 1}

\item{que.res}{the optimal clustering resolution of dataset 2}

\item{ref.norm}{whether dataset 1 is normalized}

\item{que.norm}{whether dataset 2 is normalized}

\item{dim.pca}{number of PCs to calculate (50 by default)}

\item{dim.cca}{number of canonical vectors to calculate (30 by default)}

\item{min.cluster.cells}{the minimum number of cells in a cluster}

\item{random_PCC}{the minimum value of random noise}

\item{distance_diff}{regulate the different clusters distance}

\item{distance_same}{regulate the same clusters distance}
}
\value{
a list containing the following elements:
\itemize{
\item D1_reference: SeuratObject of dataset 1
\item D2_query: SeuratObject of dataset 2
\item Matching_matrix: matching matrix, 2 is mutually most correlated clusters, 1 is mutually correlated clusters
\item cell_embedding: cell embeddings
}
}
\description{
Integrate scRNA-seq datasets
}
