% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMatch_UMAP.R
\name{ClusterMatch_UMAP}
\alias{ClusterMatch_UMAP}
\title{UMAP visualization}
\usage{
ClusterMatch_UMAP(embedding, celltype)
}
\arguments{
\item{embedding}{cell embeddings}

\item{celltype}{a data.frame, the first column is the cell and the rest are the cell's label}
}
\value{
a data.frame, the first column is the cell, the second and third columns are UMAP1 and UMAP2, and the rest are the cell's label
}
\description{
UMAP visualization
}
