% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMatch_integration.R,
%   R/ClusterMatch_matching.R, R/ClusterMatch_resolution.R
\name{Match}
\alias{Match}
\title{Matching clusters of two datasets}
\usage{
Match(data, top = 3, threshold = 1.96)

Match(data, top = 3, threshold = 1.96)

Match(data, top = 3, threshold = 1.96)
}
\arguments{
\item{data}{the correlation between clusters of two datasets}

\item{top}{the maximum number of matches for each cluster}

\item{threshold}{the minimum threshold for matching}
}
\value{
matching matrix

matching matrix

matching matrix
}
\description{
Matching clusters of two datasets

Matching clusters of two datasets

Matching clusters of two datasets
}
