% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterMatch_annotation.R
\name{Match_annotation}
\alias{Match_annotation}
\title{Matching clusters and cell types}
\usage{
Match_annotation(data, threshold = 0)
}
\arguments{
\item{data}{the correlation between clusters and cell types}

\item{threshold}{the minimum threshold for matching}
}
\value{
the matching matrix of clusters and cell types
}
\description{
Matching clusters and cell types
}
